[Pipelines] Fix test_huggingface_utilities.py tests (#60840)
Fix test_huggingface_utilities.py tests

MAX_INTEGRATION_TESTS_REV_ID: 2c220536a14e4787b151550280ec81e9600e18b3

+++ max/tests/integration/pipelines/python/test_common/pipeline_model_dummy.py
@@ -298,7 +298,7 @@ DUMMY_ARCH = SupportedArchitecture(
     name="LlamaForCausalLM",
     task=PipelineTask.TEXT_GENERATION,
     example_repo_ids=["modularai/llama-3.1"],
-    default_encoding=SupportedEncoding.float32,
+    default_encoding=SupportedEncoding.bfloat16,
     supported_encodings={
         SupportedEncoding.float32: [KVCacheStrategy.CONTINUOUS],
         SupportedEncoding.bfloat16: [KVCacheStrategy.CONTINUOUS],
